app/requirements.txt
fastapi==0.73.0
numpy==1.22.1
pandas==1.1.5
pydantic==1.8.2
python-dateutil==2.8.2
PyYAML==6.0
urllib3==1.26.7
uvicorn==0.15.0
sqlalchemy==2.0.9
psycopg2==2.9.6

